+++ compress/gzip/gzip.go
@@ -11,6 +11,10 @@ import (
 	"io"
 )
 
+import (
+	"github.com/juju/errors"
+)
+
 // DoGzipCompresss do gzip compress
 func DoGzipCompress(src []byte) []byte {
 	var in bytes.Buffer
@@ -21,10 +25,13 @@ func DoGzipCompress(src []byte) []byte {
 }
 
 // DoGzipUncompresss do gzip uncompress
-func DoGzipUncompress(compressSrc []byte) []byte {
+func DoGzipUncompress(compressSrc []byte) ([]byte, error) {
 	b := bytes.NewReader(compressSrc)
 	var out bytes.Buffer
-	r, _ := gzip.NewReader(b)
+	r, err := gzip.NewReader(b)
+	if err != nil {
+		return nil, errors.Annotate(err, "fialed to ungzip")
+	}
 	io.Copy(&out, r)
-	return out.Bytes()
+	return out.Bytes(), nil
 }