+++ compress/zlib/zip_test.go
@@ -5,8 +5,8 @@ import "testing"
 func TestDoZlibCompress(t *testing.T) {
 	var str = "hello, world"
 	zipStr := DoZlibCompress([]byte(str))
-	uncompressString := DoZlibUncompress(zipStr)
-	if string(uncompressString) != str {
-		t.Errorf("str:%q, uncompress string:%q", str, string(uncompressString))
+	uncompressString, err := DoZlibUncompress(zipStr)
+	if string(uncompressString) != str || err != nil {
+		t.Errorf("str:%q, uncompress string:%q, err:%s", str, string(uncompressString), err)
 	}
 }